refactor(#682): give close-on-outside-click one owner before a second user
#682's radio picker is the second rail popover that must dismiss when the
operator clicks past it. RailActions (#500) had the only implementation, so
this lifts it to a shared verb rather than pasting it with tweaks.

Nothing about the behaviour moves. The reasoning travels with the code and
now lives on the verb, where the next caller will read it: this is a
NON-BLOCKING document listener and deliberately not a backdrop scrim, because
a scrim swallows the click that dismissed it — with one, selecting a sidebar
channel while a popover is open costs two gestures instead of one. Capture
phase so a stopPropagation() in the tree cannot hide the click; registered
only while open, so the opening gesture cannot dismiss what it just opened.

Verified as behaviour-preserving rather than assumed: the rail suites stay
green (RailActions 58, railWhois 20, railExpandedHome 4, railMenuSafeArea 4),
and among them RailActions carries an explicit "an outside pointerdown closes
the menu without firing any action" case — so the moved behaviour is one the
existing tests actually exercise, not merely one they happen not to break.
